Begin split of server files for testing
