+++ package.json
@@ -27,7 +27,8 @@
     "mysql": "^2.18.1",
     "mysql2": "^2.2.5",
     "nodemon": "^2.0.4",
-    "sequelize": "^6.3.5"
+    "sequelize": "^6.3.5",
+    "supertest": "^5.0.0"
   },
   "devDependencies": {
     "@babel/cli": "^7.11.6",
